ASYLUM: Fixed some warnings.
git-svn-id: http://asylumengine.googlecode.com/svn/trunk@359 0bfb4aae-4ea4-11de-8d8d-752d95cf3e3c

+++ engines/asylum/actor.h
@@ -168,9 +168,9 @@ public:
 	void setWalkArea(ActionArea *target);
 	void setAction(int action);
 	void setActionByIndex(int index);
-	void drawActorAt(uint16 curX, uint16 curY);
+	void drawActorAt(uint32 curX, uint32 curY);
 	void drawActor();
-	void walkTo(uint16 curX, uint16 curY);
+	void walkTo(uint32 curX, uint32 curY);
 
 
 	int currentAction; // TODO depreciate
